requirements.txt
rembg[gpu]==2.0.74
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow>=10.0.0
